% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{count_nodes_cpp}
\alias{count_nodes_cpp}
\title{Detect Significant Physical Features (Hysteresis)}
\usage{
count_nodes_cpp(sub_df, thresh_vec)
}
\arguments{
\item{sub_df}{DataFrame with x (coordinate) and y (density)}

\item{thresh_vec}{Vector of adaptive thresholds matching sub_df rows}
}
\description{
Detect Significant Physical Features (Hysteresis)
}
