% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{erase_by_uncertainty}
\alias{erase_by_uncertainty}
\title{Erase microstate information by uncertainty threshold}
\usage{
erase_by_uncertainty(x, uncertainty)
}
\arguments{
\item{x}{A numeric vector of microstates to erase.}

\item{uncertainty}{The maximum allowable physical fluctuation.}
}
\value{
A data frame with physical and algorithmic information properties.
}
\description{
Implements Landauer erasure by mapping microstates to macrostates
within a specified physical uncertainty (x_0).
}
