% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{erase_by_uncertainty_and_depth}
\alias{erase_by_uncertainty_and_depth}
\title{Erase microstate information by both uncertainty and depth}
\usage{
erase_by_uncertainty_and_depth(x, uncertainty, depth)
}
\arguments{
\item{x}{A numeric vector of microstates to erase.}

\item{uncertainty}{The maximum allowable physical fluctuation.}

\item{depth}{The maximum target Kolmogorov complexity.}
}
\value{
A data frame with physical and algorithmic information properties.
}
\description{
Erase microstate information by both uncertainty and depth
}
