% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_nodes_cpp}
\alias{find_nodes_cpp}
\title{Detect Oscillatory Nodes in Physical State Density}
\usage{
find_nodes_cpp(sub_df, thresh, global_h_range)
}
\arguments{
\item{sub_df}{A DataFrame containing 'x' and 'y'.}

\item{thresh}{The sensitivity threshold.}

\item{global_h_range}{The total range of the histogram counts.}
}
\value{
A DataFrame of detected node coordinates (x, y).
}
\description{
Detect Oscillatory Nodes in Physical State Density
}
