% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pib_erasures}
\alias{pib_erasures}
\title{Run Stern-Brocot Particle in a Box Simulation}
\usage{
pib_erasures(momenta, dir, n_threads = 0L)
}
\arguments{
\item{momenta}{Vector of momentum values (P) to simulate.}

\item{dir}{Output directory.}

\item{n_threads}{Number of threads (0 = auto).}
}
\description{
Run Stern-Brocot Particle in a Box Simulation
}
