% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_causal_chain.R
\name{plot_causal_chain}
\alias{plot_causal_chain}
\title{Render the Complete Causal Chain Erasure Grid}
\usage{
plot_causal_chain(
  dt_meta,
  dir_raw,
  dir_densities,
  style = c("manuscript", "readme"),
  base_font = ""
)
}
\arguments{
\item{dt_meta}{data.table containing normalized_momentum and action_A.}

\item{dir_raw}{Directory containing raw erasures.}

\item{dir_densities}{Directory containing erasure densities.}

\item{style}{String, either "manuscript" or "readme".}

\item{base_font}{String for the font family.}
}
\description{
Render the Complete Causal Chain Erasure Grid
}
