% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_erasure_action.R
\name{plot_erasure_action}
\alias{plot_erasure_action}
\title{Render the Landauer Action vs Classical Action Plot}
\usage{
plot_erasure_action(
  dt_summary,
  max_action = 50,
  style = c("manuscript", "readme"),
  base_font = ""
)
}
\arguments{
\item{dt_summary}{data.table containing sequence length summary metrics.}

\item{max_action}{Numeric, maximum action to plot on the x-axis. Defaults to 50.}

\item{style}{String, either "manuscript" or "readme".}

\item{base_font}{String for the font family.}
}
\description{
Render the Landauer Action vs Classical Action Plot
}
