% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter_grid.R
\name{plot_scatter_grid}
\alias{plot_scatter_grid}
\title{Render the Microstate and Erasure Displacement Scatter Grid}
\usage{
plot_scatter_grid(
  dt_meta,
  dir_raw,
  style = c("manuscript", "readme"),
  base_font = "",
  overarching_title = NULL
)
}
\arguments{
\item{dt_meta}{data.table containing normalized_momentum and action_A.}

\item{dir_raw}{Directory containing raw erasures.}

\item{style}{String, either "manuscript" or "readme".}

\item{base_font}{String for the font family.}

\item{overarching_title}{Optional title for the readme style.}
}
\description{
Render the Microstate and Erasure Displacement Scatter Grid
}
