% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_computer.R
\name{process_action_densities}
\alias{process_action_densities}
\title{Process Densities for Multiple Target Columns}
\usage{
process_action_densities(f, out_path, target_cols)
}
\arguments{
\item{f}{A raw simulation file path.}

\item{out_path}{The output directory.}

\item{target_cols}{A character vector of column names to process.}
}
\value{
A data.table indicating success
}
\description{
Process Densities for Multiple Target Columns
}
