% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observed_state_density_computer.R
\name{process_observed_state_density}
\alias{process_observed_state_density}
\title{Process Erasure Distance Density}
\usage{
process_observed_state_density(f, out_path)
}
\arguments{
\item{f}{A file}

\item{out_path}{A dir}
}
\value{
A data.table
}
\description{
Process
}
