% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SternBrocotPhysics-package.R
\docType{package}
\name{SternBrocotPhysics-package}
\alias{SternBrocotPhysics}
\alias{SternBrocotPhysics-package}
\title{SternBrocotPhysics: Stern-Brocot Physics}
\description{
A platform for computational experiments in classical and quantum physics.
}
\author{
\strong{Maintainer}: Brian Mulloy \email{brian@homeymusic.com}

}
\keyword{internal}
