% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_counter.R
\name{count_nodes}
\alias{count_nodes}
\title{Count Nodes using Outward-Scanning Adaptive Hysteresis}
\usage{
count_nodes(density, debug = FALSE)
}
\arguments{
\item{density}{A data.table containing 'x' and 'y' columns.}

\item{debug}{Logical. If TRUE, prints decision logic to the console.}
}
\description{
Count Nodes using Outward-Scanning Adaptive Hysteresis
}
