% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_counter.R
\name{count_nodes}
\alias{count_nodes}
\title{Count Nodes using Gabor Hysteresis}
\usage{
count_nodes(density)
}
\arguments{
\item{density}{A data.table containing 'x' and 'y' columns.}
}
\description{
Count Nodes using Gabor Hysteresis
}
