% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{erase_by_depth}
\alias{erase_by_depth}
\title{Erase microstate information by a specific tree depth}
\usage{
erase_by_depth(x, depth)
}
\arguments{
\item{x}{A numeric vector of microstates to erase.}

\item{depth}{The target Kolmogorov complexity (path length).}
}
\value{
A data frame with physical and algorithmic information properties.
}
\description{
Maps microstates to macrostates by enforcing a fixed Kolmogorov complexity.
}
